Final lint warnings, test fixe
